feat(reader): Bible Book & Chapter Picker (#2)
* fix: Fix bug in caching where book chapters were not serialized

Inside of BibleBook, `chapters` was marked as `@Transient` so when it
was encoded to JSON for caching, it was being exlcuded. This caused
chapters to not load when rendering all books and chapters in the
References screen

* feat: Render Bible Books & Chapters for Selection

Add contents, viewmodel, and view data class to render all books in a
bible version and all chapters in a book. Expand and collapse each
book's chapters when a book is clicked.

* feat: Render new BibleReference when one has been selected

When a Chapter from a Book has been selected, automatically dismiss the
ReferencesScreen and begin to load and display the newly selected
reference

* fix: Failing test with new caching strategy

When BibleBook's chapters property was moved from @Transient to
@SerialName, the API response no longer matched, so the test was
failing. This fixes by changing the name to something the API won't
return but instead will be only used locally

* fix(reader): Added opt in APIs for ReferencesScreen

* feat(reader): Automatically scroll to the current reference

* fix(reader): Be smarter about when to scroll to the selected item